not going to work
I'm going to try a different tack.
I don't think the state machine approach of looking at the incoming data sequentially can work, especially since I want to have read enabled all the time to avoid the two cycle latency at the start of the read cycle.

I'm going to be processing a lot of data through the FIFO, for it to be free running the logic has to be fast and robust. It must have some way of identifying the beginning and end of data frames.